Created tests for adoption events & registration  models. #23
